refactor(jobs): change verify trusty code to actual
