+++ seahorn/jobs/storage_ipc_msg_buffer/main.c
@@ -5,62 +5,23 @@
 #include <trusty_log.h>
 #include <uapi/err.h>
 
-#include "ipc.h"
+#include <ipc.h>
 
 #include "tipc_limits.h"
 #include <interface/storage/storage.h>
 
-#include "handle_table.h"
-#include "seahorn/seahorn.h"
-
-/** entry point in ipc.c for even handling */
-extern void dispatch_event(const uevent_t *ev);
-
-static void sea_ipc_disconnect_handler(struct ipc_channel_context *context) {
-  if (context)
-    free(context);
-}
-
-static int sea_ipc_msg_handler(struct ipc_channel_context *context, void *msg,
-                               size_t msg_size) {
-  // sassert(msg_size <= MSG_BUF_MAX_SIZE);
-  struct iovec iov = {
-      .iov_base = msg,
-      .iov_len = msg_size,
-  };
-  ipc_msg_t i_msg = {
-      .iov = &iov,
-      .num_iov = 1,
-  };
-  int rc = send_msg(context->common.handle, &i_msg);
-  if (rc < 0) {
-    return rc;
-  }
-  return NO_ERROR;
-}
-
-/*
- * directly return a channel context given uuid and chan handle
- */
-static struct ipc_channel_context *
-sea_channel_connect(struct ipc_port_context *parent_ctx,
-                    const uuid_t *peer_uuid, handle_t chan_handle) {
-  struct ipc_channel_context *pctx = malloc(sizeof(struct ipc_channel_context));
-  pctx->ops.on_disconnect = sea_ipc_disconnect_handler;
-  pctx->ops.on_handle_msg = sea_ipc_msg_handler;
-  return pctx;
-}
+#include <handle_table.h>
+#include <seahorn/seahorn.h>
+#include <ipc_helper.h>
 
 /**
    verification entry point
  */
 int main(void) {
   handle_table_init(INVALID_IPC_HANDLE, INVALID_IPC_HANDLE, INVALID_IPC_HANDLE);
-  struct ipc_port_context ctx = {
-      .ops = {.on_connect = sea_channel_connect},
-  };
+  struct ipc_port_context* ctx = create_port_context();
   int rc =
-      ipc_port_create(&ctx, STORAGE_DISK_PROXY_PORT, 1, STORAGE_MAX_BUFFER_SIZE,
+      ipc_port_create(ctx, STORAGE_DISK_PROXY_PORT, 1, STORAGE_MAX_BUFFER_SIZE,
                       IPC_PORT_ALLOW_TA_CONNECT | IPC_PORT_ALLOW_NS_CONNECT);
 
   if (rc < 0) {
@@ -78,7 +39,7 @@ int main(void) {
     return rc;
   }
   if (rc == NO_ERROR) { /* got an event */
-    dispatch_event(&event1);
+    sea_dispatch_event(&event1);
   }
   // get second event, could be either port or channel
   uevent_t event2;
@@ -90,10 +51,10 @@ int main(void) {
     return rc;
   }
   if (rc == NO_ERROR) { /* got an event */
-    dispatch_event(&event2);
+    sea_dispatch_event(&event2);
   }
 
-  ipc_port_destroy(&ctx);
+  ipc_port_destroy(ctx);
 
   return 0;
 }
